Read Million Song Dataset's path from env-file
Added the .gitignore-file for the project to ignore the .env-file that can be locally used. Then implemented a functionality to read an enviroment variable MSD_PATH to be given for a function called initialize_MSD_file_selector which creates a Panel FileSelector widget with root directory pointing to MSD_PATH or to '~'.