Refactoring and changes...
Changed functionality so that when a user selects a song then automativally recommendations are calculated and outputs are showed. See that currently selecting more than one song produces nothing because giving multiple songs for the model is not yet implemented. These changes led to major refactoring and app.py is much simpler yet more easy and efficient.